+++ src/actions/actionTypes.ts
@@ -60,7 +60,7 @@ export const FINISH_INVITE = 'FINISH_INVITE';
 export const SET_EDITOR_SESSION_ID = 'SET_EDITOR_SESSION_ID';
 export const SET_WEBSOCKET_STATUS = 'SET_WEBSOCKET_STATUS';
 
-/** Materials */
+/** Material */
 export const CREATE_MATERIAL_FOLDER = 'CREATE_MATERIAL_FOLDER';
 export const DELETE_MATERIAL = 'DELETE_MATERIAL';
 export const DELETE_MATERIAL_FOLDER = 'DELETE_MATERIAL_FOLDER';
