+++ src/actions/index.ts
@@ -2,7 +2,7 @@ export * from './collabEditing';
 export * from './commons';
 export * from './game';
 export * from './interpreter';
-export * from './materials';
+export * from './material';
 export * from './playground';
 export * from './session';
 export * from './workspaces';
