+++ src/styles/_material.scss
@@ -1,0 +1,40 @@
+.Input {
+  width: 90%;
+  outline: none;
+  resize: none;
+  display: flex;
+  justify-content: flex-start;
+  margin: 0 auto;
+}
+
+.Material {
+  /* To show the table borders that get cut out by a Card */
+  padding: 1px;
+
+  /** 
+   *  To prevent misalignment of column header and content 
+   *  for large displays. This effectively flushes the content left.
+   *  See https://github.com/source-academy/cadet-frontend/pull/148
+   */
+  .ag-body-container {
+    display: block;
+    /* Override ag-grid style to suppress unnecessary vertical scrollbar (bugfix) */
+    margin-bottom: 0;
+  }
+
+  .#{$ns}-icon {
+    /* Vertically centre appearance of blueprintjs icons in ag-grid */
+    margin-top: 6px;
+  }
+
+  .ag-paging-panel {
+    border-top: none;
+    font-weight: bold;
+    padding: 10px;
+  }
+
+  :focus {
+    /* Disable rendering blue border when a sub-component in grid is in focus */
+    outline: none;
+  }
+}
