+++ src/styles/index.scss
@@ -24,7 +24,7 @@
 @import 'dropdown';
 @import 'game';
 @import 'login';
-@import 'materials';
+@import 'material';
 @import 'navigationBar';
 @import 'notfound';
 @import 'playground';
